projectile-indexing-method
Use the native method, though much more slower the first time, we can
globally set ignored dirs or files, and project specific ignored dirs
or files via the .projectile file. Actually the other fast index
methods should take the settings and .projectile into account, or we
can put the dirs and files which are not want to be cached in the
.gitignore but don't commit the changes, this sounds not like a good
way.